feat(alerts): unify the add-alert UI and surface active alerts in the drawer
A single "Add alert" entry point now serves both alert shapes: the new
AlertDrawer has a type toggle (item price / retainer undercut) that swaps
the form fields, replacing the separate CreateAlertDrawer and
UndercutAlertDrawer components. The same drawer lists the user's existing
alerts of the selected type with a remove button, so the modal on
/retainers/undercuts (which opens the drawer defaulted to the undercut
type) can now unsubscribe without a trip back to /alerts.

The four near-verbatim copies of the "Deliver to" endpoint checkbox list
collapse into a shared EndpointPicker used by AlertDrawer,
AlertConfigDrawer, and ListSubscribeDrawer.

While touching AlertRulesPanel, its hardcoded toast/empty-state/status
strings now use the locale keys that already existed for them
(alerts_alert_enabled/disabled/deleted, alerts_empty_state,
alerts_status_enabled/disabled). New drawer strings are keyed in all
seven locales; the four keys orphaned by the deleted drawers are removed.

Fixes #1065
Fixes #1066
