+++ ultros-frontend/ultros-app/locales/cn.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "尚无通知端点。",
     "alert_drawer_no_endpoints_link": "添加一个",
     "alert_drawer_no_endpoints_suffix": " 后再创建提醒。",
+    "add_alert_button": "添加提醒",
+    "alert_kind_label": "提醒类型",
+    "alert_kind_item_price": "物品价格",
+    "alert_kind_undercut": "雇员被低价压过",
+    "alert_drawer_active_heading": "当前提醒",
+    "alert_drawer_active_empty": "暂无此类型的提醒。",
+    "alert_drawer_threshold_below": "≤ {{price}} 金币",
     "alerts_tab_endpoints": "通知端点",
     "alerts_tab_rules": "提醒规则",
     "alerts_tab_history": "历史",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "没有你既能管理、Bot 又能在频道中写入的共同 Discord 服务器。",
     "endpoints_channel_label": "频道",
     "endpoints_create_button": "创建",
-    "undercut_alert_open_button": "被低价压过提醒",
-    "undercut_alert_title": "创建被低价压过提醒",
     "undercut_alert_description": "当其他挂单低于你认领的雇员价格时通知你。",
     "undercut_alert_margin_label": "利润率百分比",
     "undercut_alert_err_margin_number": "利润率必须为数字",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "雇员被低价压过",
     "alerts_margin_percent": "{{margin}}% 利润率",
     "alerts_list_update_rule": "清单更新",
-    "create_alert_button": "创建提醒",
-    "create_alert_title": "创建价格提醒",
     "create_alert_item_label": "物品",
     "create_alert_search_placeholder": "搜索物品...",
     "create_alert_change_item": "更改",
