+++ ultros-frontend/ultros-app/locales/de.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "Noch keine Endpunkte. ",
     "alert_drawer_no_endpoints_link": "Einen hinzufügen",
     "alert_drawer_no_endpoints_suffix": " bevor du Alarme erstellst.",
+    "add_alert_button": "Benachrichtigung hinzufügen",
+    "alert_kind_label": "Benachrichtigungstyp",
+    "alert_kind_item_price": "Gegenstandspreis",
+    "alert_kind_undercut": "Gehilfen-Unterbietung",
+    "alert_drawer_active_heading": "Deine aktiven Benachrichtigungen",
+    "alert_drawer_active_empty": "Noch keine Benachrichtigungen dieses Typs.",
+    "alert_drawer_threshold_below": "≤ {{price}} Gil",
     "alerts_tab_endpoints": "Endpunkte",
     "alerts_tab_rules": "Alarmregeln",
     "alerts_tab_history": "Verlauf",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "Keine gemeinsamen Discord-Server, in denen du den Server verwalten kannst und der Bot in einen Kanal schreiben darf.",
     "endpoints_channel_label": "Kanal",
     "endpoints_create_button": "Erstellen",
-    "undercut_alert_open_button": "Unterbietungs-Alarm",
-    "undercut_alert_title": "Unterbietungs-Alarm erstellen",
     "undercut_alert_description": "Lass dich benachrichtigen, wenn ein anderes Angebot einen deiner beanspruchten Gehilfen unterbietet.",
     "undercut_alert_margin_label": "Margen-Prozentsatz",
     "undercut_alert_err_margin_number": "Marge muss eine Zahl sein",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "Gehilfen-Unterbietungen",
     "alerts_margin_percent": "{{margin}} % Marge",
     "alerts_list_update_rule": "Listen-Updates",
-    "create_alert_button": "Alarm erstellen",
-    "create_alert_title": "Preisalarm erstellen",
     "create_alert_item_label": "Gegenstand",
     "create_alert_search_placeholder": "Items suchen...",
     "create_alert_change_item": "Ändern",
