+++ ultros-frontend/ultros-app/locales/en.json
@@ -1119,6 +1119,13 @@
     "alert_drawer_no_endpoints_prefix": "No endpoints yet. ",
     "alert_drawer_no_endpoints_link": "Add one",
     "alert_drawer_no_endpoints_suffix": " before creating alerts.",
+    "add_alert_button": "Add alert",
+    "alert_kind_label": "Alert type",
+    "alert_kind_item_price": "Item price",
+    "alert_kind_undercut": "Retainer undercut",
+    "alert_drawer_active_heading": "Your active alerts",
+    "alert_drawer_active_empty": "No alerts of this type yet.",
+    "alert_drawer_threshold_below": "≤ {{price}} gil",
     "alerts_tab_endpoints": "Endpoints",
     "alerts_tab_rules": "Alert rules",
     "alerts_tab_history": "History",
@@ -1170,8 +1177,6 @@
     "endpoints_no_discord_servers": "No shared Discord servers where you can manage the server and the bot can write to a channel.",
     "endpoints_channel_label": "Channel",
     "endpoints_create_button": "Create",
-    "undercut_alert_open_button": "Undercut alert",
-    "undercut_alert_title": "Create undercut alert",
     "undercut_alert_description": "Get notified when another listing undercuts one of your claimed retainers.",
     "undercut_alert_margin_label": "Margin percent",
     "undercut_alert_err_margin_number": "Margin must be a number",
@@ -1183,8 +1188,6 @@
     "alerts_retainer_undercut_rule": "Retainer undercuts",
     "alerts_margin_percent": "{{margin}}% margin",
     "alerts_list_update_rule": "list updates",
-    "create_alert_button": "Create alert",
-    "create_alert_title": "Create price alert",
     "create_alert_item_label": "Item",
     "create_alert_search_placeholder": "Search items...",
     "create_alert_change_item": "Change",
