+++ ultros-frontend/ultros-app/locales/fr.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "Aucun point de distribution. ",
     "alert_drawer_no_endpoints_link": "En ajouter un",
     "alert_drawer_no_endpoints_suffix": " avant de créer des alertes.",
+    "add_alert_button": "Ajouter une alerte",
+    "alert_kind_label": "Type d'alerte",
+    "alert_kind_item_price": "Prix d'un objet",
+    "alert_kind_undercut": "Sous-cotation de serviteur",
+    "alert_drawer_active_heading": "Vos alertes actives",
+    "alert_drawer_active_empty": "Aucune alerte de ce type pour le moment.",
+    "alert_drawer_threshold_below": "≤ {{price}} gils",
     "alerts_tab_endpoints": "Points de distribution",
     "alerts_tab_rules": "Règles d'alerte",
     "alerts_tab_history": "Historique",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "Aucun serveur Discord partagé où vous pouvez gérer le serveur et où le bot peut écrire dans un salon.",
     "endpoints_channel_label": "Salon",
     "endpoints_create_button": "Créer",
-    "undercut_alert_open_button": "Alerte de sous-cotation",
-    "undercut_alert_title": "Créer une alerte de sous-cotation",
     "undercut_alert_description": "Soyez notifié lorsqu'une autre annonce sous-cote l'un de vos serviteurs revendiqués.",
     "undercut_alert_margin_label": "Pourcentage de marge",
     "undercut_alert_err_margin_number": "La marge doit être un nombre",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "Sous-cotations de serviteur",
     "alerts_margin_percent": "{{margin}}% de marge",
     "alerts_list_update_rule": "mises à jour de liste",
-    "create_alert_button": "Créer une alerte",
-    "create_alert_title": "Créer une alerte de prix",
     "create_alert_item_label": "Objet",
     "create_alert_search_placeholder": "Rechercher des objets...",
     "create_alert_change_item": "Modifier",
