+++ ultros-frontend/ultros-app/locales/ja.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "配信先がまだありません。",
     "alert_drawer_no_endpoints_link": "追加する",
     "alert_drawer_no_endpoints_suffix": " アラートを作成する前に。",
+    "add_alert_button": "アラートを追加",
+    "alert_kind_label": "アラートの種類",
+    "alert_kind_item_price": "アイテム価格",
+    "alert_kind_undercut": "雇員のアンダーカット",
+    "alert_drawer_active_heading": "有効なアラート",
+    "alert_drawer_active_empty": "この種類のアラートはまだありません。",
+    "alert_drawer_threshold_below": "≤ {{price}} gil",
     "alerts_tab_endpoints": "配信先",
     "alerts_tab_rules": "アラートルール",
     "alerts_tab_history": "履歴",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "サーバー管理権限があり、かつボットがチャンネルに書き込める共通のDiscordサーバーがありません。",
     "endpoints_channel_label": "チャンネル",
     "endpoints_create_button": "作成",
-    "undercut_alert_open_button": "アンダーカットアラート",
-    "undercut_alert_title": "アンダーカットアラートを作成",
     "undercut_alert_description": "あなたの登録雇員の出品が他の出品にアンダーカットされたときに通知します。",
     "undercut_alert_margin_label": "マージン (%)",
     "undercut_alert_err_margin_number": "マージンは数値である必要があります",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "雇員のアンダーカット",
     "alerts_margin_percent": "{{margin}}% マージン",
     "alerts_list_update_rule": "リストの更新",
-    "create_alert_button": "アラート作成",
-    "create_alert_title": "価格アラートを作成",
     "create_alert_item_label": "アイテム",
     "create_alert_search_placeholder": "アイテムを検索...",
     "create_alert_change_item": "変更",
