+++ ultros-frontend/ultros-app/locales/ko.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "엔드포인트가 아직 없습니다. ",
     "alert_drawer_no_endpoints_link": "추가하기",
     "alert_drawer_no_endpoints_suffix": " 알림을 만들기 전에.",
+    "add_alert_button": "알림 추가",
+    "alert_kind_label": "알림 유형",
+    "alert_kind_item_price": "아이템 가격",
+    "alert_kind_undercut": "모험가 가격 인하",
+    "alert_drawer_active_heading": "활성 알림",
+    "alert_drawer_active_empty": "이 유형의 알림이 아직 없습니다.",
+    "alert_drawer_threshold_below": "≤ {{price}} 길",
     "alerts_tab_endpoints": "엔드포인트",
     "alerts_tab_rules": "알림 규칙",
     "alerts_tab_history": "기록",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "관리 권한이 있으면서 봇이 채널에 쓸 수 있는 공통 디스코드 서버가 없습니다.",
     "endpoints_channel_label": "채널",
     "endpoints_create_button": "생성",
-    "undercut_alert_open_button": "가격 인하 알림",
-    "undercut_alert_title": "가격 인하 알림 만들기",
     "undercut_alert_description": "다른 매물이 등록된 모험가 가격보다 낮게 올라오면 알려드립니다.",
     "undercut_alert_margin_label": "마진 퍼센트",
     "undercut_alert_err_margin_number": "마진은 숫자여야 합니다",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "모험가 가격 인하 감지",
     "alerts_margin_percent": "{{margin}}% 마진",
     "alerts_list_update_rule": "리스트 업데이트",
-    "create_alert_button": "알림 생성",
-    "create_alert_title": "가격 알림 생성",
     "create_alert_item_label": "아이템",
     "create_alert_search_placeholder": "아이템 검색...",
     "create_alert_change_item": "변경",
