+++ ultros-frontend/ultros-app/locales/tc.json
@@ -1106,6 +1106,13 @@
     "alert_drawer_no_endpoints_prefix": "尚無通知端點。",
     "alert_drawer_no_endpoints_link": "新增一個",
     "alert_drawer_no_endpoints_suffix": " 後再建立提醒。",
+    "add_alert_button": "新增提醒",
+    "alert_kind_label": "提醒類型",
+    "alert_kind_item_price": "物品價格",
+    "alert_kind_undercut": "雇員被低價壓過",
+    "alert_drawer_active_heading": "目前提醒",
+    "alert_drawer_active_empty": "尚無此類型的提醒。",
+    "alert_drawer_threshold_below": "≤ {{price}} gil",
     "alerts_tab_endpoints": "通知端點",
     "alerts_tab_rules": "提醒規則",
     "alerts_tab_history": "歷史",
@@ -1167,8 +1174,6 @@
     "endpoints_no_discord_servers": "沒有你既能管理、Bot 又能在頻道中寫入的共同 Discord 伺服器。",
     "endpoints_channel_label": "頻道",
     "endpoints_create_button": "建立",
-    "undercut_alert_open_button": "被低價壓過提醒",
-    "undercut_alert_title": "建立被低價壓過提醒",
     "undercut_alert_description": "當其他掛單低於你認領的雇員價格時通知你。",
     "undercut_alert_margin_label": "利潤率百分比",
     "undercut_alert_err_margin_number": "利潤率必須為數字",
@@ -1180,8 +1185,6 @@
     "alerts_retainer_undercut_rule": "雇員被低價壓過",
     "alerts_margin_percent": "{{margin}}% 利潤率",
     "alerts_list_update_rule": "清單更新",
-    "create_alert_button": "建立提醒",
-    "create_alert_title": "建立價格提醒",
     "create_alert_item_label": "物品",
     "create_alert_search_placeholder": "搜尋物品...",
     "create_alert_change_item": "更改",
