+++ ultros-frontend/ultros-app/src/components/mod.rs
@@ -5,6 +5,7 @@ pub mod add_recipe_to_list;
 pub mod add_set_to_list;
 pub mod add_to_list;
 pub mod alert_config_drawer;
+pub mod alert_drawer;
 pub mod alert_rules_panel;
 pub mod app_shell;
 pub mod chart_toolbar;
@@ -13,8 +14,8 @@ pub mod clipboard;
 pub mod confidence_badge;
 pub mod crafter_settings;
 pub mod crafting_cost;
-pub mod create_alert_drawer;
 pub mod datacenter_name;
+pub(crate) mod endpoint_picker;
 pub mod endpoints_panel;
 pub mod filter_card;
 pub mod filter_chip;
@@ -76,7 +77,6 @@ pub mod toolbar;
 pub mod tooltip;
 pub mod top_opportunity;
 pub mod ui_text;
-pub mod undercut_alert_drawer;
 pub mod virtual_scroller;
 pub mod wasm_loading_indicator;
 pub mod world_name;
